+++ tests/ops/service/privacy_request/test_consent_email_batch_send.py
@@ -277,7 +277,7 @@ class TestConsentEmailBatchSend:
                 identities={"ljt_readerID": "12345"},
                 consent_preferences=[
                     Consent(
-                        data_use="advertising", data_use_description=None, opt_in=False
+                        data_use="advertising", data_use_description=None, opt_in=False, conflicts_with_gpc=False, has_gpc_flag=False,
                     )
                 ],
             ),
@@ -285,7 +285,7 @@ class TestConsentEmailBatchSend:
                 identities={"ljt_readerID": "abcde"},
                 consent_preferences=[
                     Consent(
-                        data_use="advertising", data_use_description=None, opt_in=False
+                        data_use="advertising", data_use_description=None, opt_in=False, conflicts_with_gpc=False, has_gpc_flag=False,
                     )
                 ],
             ),
@@ -387,6 +387,8 @@ class TestConsentEmailBatchSendHelperFunctions:
                     "data_use": "Advertising, Marketing or Promotion",
                     "data_use_description": None,
                     "opt_in": False,
+                    "has_gpc_flag": False,
+                    "conflicts_with_gpc": False,
                 },
             ],
         }